% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_tibble.R
\name{accounts}
\alias{accounts}
\title{Retrieve Accounts}
\usage{
accounts(token)
}
\arguments{
\item{token}{Required. Created by \code{token("accounts")}.}
}
\value{
A tbl.
}
\description{
Retrieve accounts from authenticated login used in token.
}
\examples{
tok <- token("accounts")
tbl <- accounts(tok)

}
