% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{token}
\alias{token}
\title{Create Authentication Token}
\usage{
token(scope = c("accounts", "transactions"), force_new = FALSE,
  market = "SE")
}
\arguments{
\item{scope}{A character vector. Possible scopes are \code{"user", "credentials",
"accounts", "transactions", "investments", "statistics"}. Multiple scopes
can be used in same token.}

\item{force_new}{Logical. Should a new token be requested even though cache
exist? If there has been a while since token cache was created, then this
probably needs to be \code{TRUE}.}

\item{market}{A character string. Market (country) in which bank is accessed.}
}
\value{
An httr::oauth2.0_token object.
}
\description{
Create Authentication Token
}
\examples{
token("accounts")
token(c("accounts", "transactions"))
token("accounts", force_new = TRUE)

}
