% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_tibble.R
\name{transactions}
\alias{transactions}
\title{Retrieve Transactions}
\usage{
transactions(token)
}
\arguments{
\item{token}{Required. Created by \code{token("transactions")}.}
}
\value{
A tbl.
}
\description{
Retrieve transactions from authenticated login used in token.
}
\examples{
tok <- token("transactions")
tbl <- transactions(tok)

}
