% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_tibble.R
\name{instruments}
\alias{instruments}
\title{Retrieve Instruments}
\usage{
instruments(token)
}
\arguments{
\item{token}{Required. Created by \code{token("instruments")}.}
}
\value{
A tbl.
}
\description{
Retrieve instruments from authenticated login used in token.
}
\examples{
tok <- token("instruments")
tbl <- instruments(tok)

}
