% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{token}
\alias{token}
\title{Create Authentication Token}
\usage{
token(scope = c("accounts", "transactions"), force_new = FALSE,
  market = "SE", provider = NULL, username = NULL)
}
\arguments{
\item{scope}{Required character vector. Possible scopes are \code{"accounts",
"transactions", "investments"}. Multiple scopes can be used
in same token.}

\item{force_new}{Required logical. Should new token be requested even though
cache exist? If there has been a while since token cache was created, then
this probably needs to be \code{TRUE}.}

\item{market}{Required string. Market (country) in which bank is accessed.}

\item{provider}{Optional string. Unique name of provider (eg.
\code{sbab-bankid}). If this is provided, the provider will be chosen for
the user. If not, the user can select a provider in a list.}

\item{username}{Optional string. If this is provided, the username field will be
pre-filled. If not, the user can type it in.}
}
\value{
An httr::oauth2.0_token object.
}
\description{
Create Authentication Token
}
\examples{
token("accounts")
token(c("accounts", "transactions"))
token("accounts", force_new = TRUE)

}
